// userInterface/src/types/customerProfile.type.ts
export interface CustomerProfile {
  id: string;
  firstName?: string;
  lastName?: string;
  fullName?: string;
  userName?: string;
  email?: string;
  phoneNumber?: string;
  phone?: string;
  address?: string;
  role?: string;
  createAt?: string;
  createdAt?: string;
  emailConfirmed?: boolean;
}